backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.16
pyjwt==2.8.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
asyncpg>=0.30.0